[#38] Rename: "Rooms 패키지명 변경"
