--- src/main/java/com/sixman/roomus/Rooms/command/application/dto/UpdateFunitureInfoDTO.java
@@ -1,10 +1,0 @@
-package com.sixman.roomus.Rooms.command.application.dto;
-
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-public class UpdateFunitureInfoDTO {
-    private List<AssignmentInfoDTO> datas;
-}
