--- src/main/java/com/sixman/roomus/Rooms/command/domain/exception/NotRoomOwnerException.java
@@ -1,6 +1,0 @@
-package com.sixman.roomus.Rooms.command.domain.exception;
-
-public class NotRoomOwnerException extends RuntimeException {
-    public NotRoomOwnerException(String s) {
-    }
-}
