--- src/main/java/com/sixman/roomus/Rooms/command/domain/repository/RoomRepository.java
@@ -1,7 +1,0 @@
-package com.sixman.roomus.Rooms.command.domain.repository;
-
-import com.sixman.roomus.Rooms.command.domain.model.Room;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RoomRepository extends JpaRepository<Room, Integer> {
-}
