--- src/main/java/com/sixman/roomus/Rooms/query/dto/RoomSummaryResponseDTO.java
@@ -1,31 +1,0 @@
-package com.sixman.roomus.Rooms.query.dto;
-
-import lombok.*;
-
-import java.util.Date;
-
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@ToString
-public class RoomSummaryResponseDTO {
-    private int roomNo;
-    private int memberNo;
-    private String RoomName;
-    private boolean access;
-    private String category;
-    private String description;
-    private float xsize;
-    private float ysize;
-    private float zsize;
-    private int door;
-    private Date createdDate;
-    private Date lastModifiedDate;
-    private Date deletedDate;
-    private boolean isDelete;
-    private String screenShotUrl;
-//    private List<FurnitureArrangementData> furnitureArrangementList;
-
-}
