--- src/main/java/com/sixman/roomus/Rooms/query/repository/FurnitureArrangementDataRepository.java
@@ -1,7 +1,0 @@
-package com.sixman.roomus.Rooms.query.repository;
-
-import com.sixman.roomus.Rooms.query.model.FurnitureArrangementData;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface FurnitureArrangementDataRepository extends JpaRepository<FurnitureArrangementData, Integer> {
-}
