+++ src/main/java/com/sixman/roomus/rooms/command/application/dto/UpdateFunitureInfoDTO.java
@@ -1,0 +1,10 @@
+package com.sixman.roomus.rooms.command.application.dto;
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class UpdateFunitureInfoDTO {
+    private List<AssignmentInfoDTO> datas;
+}
