+++ src/main/java/com/sixman/roomus/rooms/command/domain/exception/NotRoomOwnerException.java
@@ -1,0 +1,6 @@
+package com.sixman.roomus.rooms.command.domain.exception;
+
+public class NotRoomOwnerException extends RuntimeException {
+    public NotRoomOwnerException(String s) {
+    }
+}
