+++ src/main/java/com/sixman/roomus/rooms/command/domain/repository/RoomRepository.java
@@ -1,0 +1,7 @@
+package com.sixman.roomus.rooms.command.domain.repository;
+
+import com.sixman.roomus.rooms.command.domain.model.Room;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RoomRepository extends JpaRepository<Room, Integer> {
+}
