+++ src/main/java/com/sixman/roomus/rooms/command/domain/service/RoomStorageService.java
@@ -1,0 +1,11 @@
+package com.sixman.roomus.rooms.command.domain.service;
+
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.IOException;
+
+public interface RoomStorageService {
+
+    String fileSaveToStorage(MultipartFile file) throws IOException;
+
+}
