+++ src/main/java/com/sixman/roomus/rooms/query/repository/FurnitureArrangementDataRepository.java
@@ -1,0 +1,7 @@
+package com.sixman.roomus.rooms.query.repository;
+
+import com.sixman.roomus.rooms.query.model.FurnitureArrangementData;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface FurnitureArrangementDataRepository extends JpaRepository<FurnitureArrangementData, Integer> {
+}
